+++ lectures/c++/06_copy_move_semantics/01_surprise.cc
