+++ lectures/c++/06_copy_move_semantics/04_buggy_vector.cc
@@ -1,14 +1,15 @@
 #include <iostream>
-#include <memory>
 
 template <typename num>
 class Vector {
-  std::unique_ptr<num[]> elem;
+  num* elem;
   std::size_t _size;
 
  public:
   explicit Vector(const std::size_t length)
       : elem{new num[length]{}}, _size{length} {}
+
+  ~Vector() noexcept { delete[] elem; }
 
   const num& operator[](const std::size_t& i) const noexcept { return elem[i]; }
   num& operator[](const std::size_t& i) noexcept { return elem[i]; }
@@ -16,16 +17,44 @@ class Vector {
   std::size_t size() const noexcept { return _size; }
 
   // range-for
-  const num* begin() const noexcept { return elem.get(); }
-  num* begin() noexcept { return elem.get(); }
+  const num* begin() const noexcept { return elem; }
+  num* begin() noexcept { return elem; }
 
-  const num* end() const noexcept { return elem.get() + _size; }
-  num* end() noexcept { return elem.get() + _size; }
+  const num* end() const noexcept { return elem + _size; }
+  num* end() noexcept { return elem + _size; }
 };
 
 int main() {
-  Vector<double> v1{7};
-  // Vector<double> v2{v1}; // default copy ctor calls a deleted function by
-  // std::unique_ptr
+  Vector<int> v1{3};
+  v1[0] = 1;
+  v1[1] = 2;
+  v1[2] = 3;
+
+  std::cout << "v1: ";
+  for (const auto x : v1)
+    std::cout << x << " ";
+  std::cout << std::endl;
+
+  Vector<int> v2{v1};  // default copy constructor
+
+  std::cout << "v2 after default copy ctor: ";
+  for (const auto x : v2)
+    std::cout << x << " ";
+  std::cout << std::endl;
+
+  v1[0] = 99;
+
+  std::cout << "v2 after v1 has been changed: ";
+  for (const auto x : v2)
+    std::cout << x << " ";
+  std::cout << std::endl;
+
+  v2[1] = -999;
+
+  std::cout << "v1 after v2 has been changed: ";
+  for (const auto x : v1)
+    std::cout << x << " ";
+  std::cout << std::endl;
+
   return 0;
 }
