+++ lectures/c++/06_copy_move_semantics/06_copy_move.cc
